requirements.txt
streamlit
pandas
plotly
glob2
python-dotenv
yfinance
textblob
talib
vadersentiment
beautifulsoup4
backtrader
requests
schedule
